+++ HISTORY.rst
@@ -1,6 +1,11 @@
 History
 -------
 
+next release
+^^^^^^^^^^^^
+
+* Use Jinja2 for templates
+
 0.1.2 (2011-08-05)
 ^^^^^^^^^^^^^^^^^^